Allow arbitrary strings as entity type
This change is based on the discussion in Berlin last week.
It should be possible to add new entity types in the future,
maybe even in some local configuration, and hence the types
need to be extendable. The strings used to identify the known
types are the IRIs of the corresponding classes in the RDF
export.