New classes for processing MediaWiki XML dumps
The main processing class is MwDumpFileFormatProcessorImpl, which reads
XML to extract revisions of pages. It forwards these revisions to a
class that implements MwRevisionProcessor. Two such classes exist:
MwRevisionProcessorBroker receives revisions and distributes them to
further MwRevisionProcessors that subscribe to specific revisions (e.g.,
only the most current revisions of type "wikidata-item"); the other
class is StatisticsMwRevisionProcessor, which gathers basic statistics
about the revisions it receives and logs them.