Change "base IRI" of entity id values to "site IRI"
* Site IRI is the same string as before, but now interpreted as an
identifier for the site the entity comes from.
* A new getter makes this string retrievable (extended interface of
EntityIDValue).

This is necessary to ensure that one can construct new ItenIdValues and
PropertyIdValues from the information that one can get from an object of
this type. The re-interpretation of "base IRI" as "site IRI" is to
prepare for future Wikibase changes where entities may have different
IDs and different ways of constructing IRIs (e.g., Commons file name IDs
and URL-encoding as part of IRI construction).