Modify Jackson's setting of site IRIs to reduce memory usage
The site IRI needs to be injected after parsing, since it is not in
JSON. The new code only stores the site IRI, not a pointer to the whole
EntityDocument. This significantly reduces memory usage in applications
that store EntityIdValues from a JSON parse in memory.

Note that it would use even less memory to convert the JSON objects to
regular WDTK objects.