Change implementations to accept arbitrary instances of the interfaces of their fields at construction.
This implies that we are now requiring all subclasses of our interfaces to be properly serializable with Jackson.
This avoids a lot of copying, and it's fairly straightforward to ensure that custom implementations comply with
this (as they only need to implement serialization).

Breaking change: the `references` field of Statement is now of type List<Reference> instead of List<? extends Reference>.
The wildcard is useless now that we do not have any copying in place.