Disable failing test testFindStatementDatatypeIdValue
This test was testing findStatementDatatypeIdValue, which is
supposed to be used when dealing with statements which have a datatype
as target value - but as far as I can tell this does not exist,
at least on Wikidata.

Maybe this is something Wikibase allows in general, hence we want to
keep it? In that case, the corresponding JSON type and IRIs should be
provided where applicable.

Otherwise, we need to clean this up:
- DatatypeIdValue should not inherit from Value
- StatementDocument should not have methods to return DatatypeIdValues from claims
- ...