Expose RDFConverter building blocks to improve extensibility
This patch factors code from RdfConverter into a new AbstractRdfConverter base class.
The advantage of that approach is that extending the RDF export becomes easier:
you can create your own subclass of the AbstractRdfConverter but still reuse all
the `writeXXXX` parts to perform the RDF generation for property/item documents
and their parts.

The current RdfConverter now makes use of that to provide the previous
interface, where control of the parts to export is guided by a set of simple flags.

Using AbstractRdfConverter, much more complex filtering can be implemented by
extending the class, for example exporting terms only for some entities.